sentinel audit/simple_requirements.txt
azure-identity>=1.12.0
azure-mgmt-securityinsight>=1.0.0
azure-core>=1.20.0
python-dotenv>=1.0.0